Add files via upload
Developed a clone of the Myntra e-commerce website, replicating the core features and user interface to provide an online shopping experience. Built the homepage of myntra using HTML and CSS for the front-end development. The homepage includes a navigation bar with categories like men, women, kids, and offers, and a prominent search bar for quick product searches. Implemented a product carousel showcasing popular items and new arrivals.